requirements.txt
ffmpeg
PyAudio
SpeechRecognition
pydub
groq 
langchain-groq
python-dotenv
gTTS
flask
whisper
ipykernel
streamlit
huggingface_hub
langchain-core
langchain_community
langchain
langchain-huggingface
sounddevice
soundfile